+++ askbot/startup_procedures.py
@@ -600,12 +600,7 @@ def test_avatar():
     """if "avatar" is in the installed apps,
     checks that the module is actually installed"""
     if 'avatar' in django_settings.INSTALLED_APPS:
-        try_import('Image', 'PIL', short_message = True)
-        try_import(
-            'avatar',
-            '-e git+git://github.com/ericflo/django-avatar.git#egg=avatar',
-            short_message = True
-        )
+        try_import('avatar', 'django-avatar', short_message=True)
 
 def test_haystack():
     if 'haystack' in django_settings.INSTALLED_APPS: