+++ askbot/views/avatar_views.py
@@ -1,202 +1,201 @@
-"""
-todo: remove this module - not needed any more
-
-this is an unfortunate copy-paste (mostly)
-from the avatar app - the reason is that django-avatar app
-does not support jinja templates
-"""
-import urllib
-from django.http import HttpResponseRedirect
-from django.template import RequestContext
-from django.template.loader import get_template
-from django.shortcuts import render
-from django.utils.translation import ugettext as _
-from django.views.decorators import csrf
-from django.conf import settings
-
-from django.contrib.auth.decorators import login_required
-
-from avatar.forms import PrimaryAvatarForm, DeleteAvatarForm, UploadAvatarForm
+from avatar.conf import settings as avatar_settings
+from avatar.forms import PrimaryAvatarForm, UploadAvatarForm
 from avatar.models import Avatar
-from avatar.settings import AVATAR_MAX_AVATARS_PER_USER, AVATAR_DEFAULT_SIZE
-from avatar.util import get_primary_avatar, get_default_avatar_url
-from avatar.views import render_primary as django_avatar_render_primary
+from avatar.signals import avatar_updated
+from askbot.models import User
+from django.contrib.auth.decorators import login_required
+from django.core.urlresolvers import reverse
+from django.http import HttpResponseRedirect
+from django.shortcuts import get_object_or_404, redirect, render
+import functools
 
-from askbot import models
+def admin_or_owner_required(func):
+    """decorator that allows admin or account owner to
+    call the view function"""
+    @functools.wraps(func)
+    def wrapped(request, user_id=None):
+        if request.user.is_authenticated():
+            if request.user.is_administrator() or request.user.id == user_id:
+                return func(request, user_id)
+        #delegate to do redirect to the login_required
+        return login_required(func)(request, user_id)
+    return wrapped
 
-notification = False
-if 'notification' in settings.INSTALLED_APPS:
-    from notification import models as notification
 
-friends = False
-if 'friends' in settings.INSTALLED_APPS:
-    friends = True
-    from friends.models import Friendship
-
-def _get_next(request):
+def get_avatar_data(user, avatar_size):
+    """avatar data and boolean, which is true is user has custom avatar
+    avatar data is a list of dictionaries, one for each avatar with keys:
+    * id - avatar id (this field is missing for gravatar and default avatar)
+    * avatar_type (string , 'uploaded_avatar', 'gravatar', 'default_avatar')
+    * url (url to avatar of requested size)
+    * is_primary (True if primary)
+    Primary avatar data must be first in the list.
+    There will always be only one primary_avatar.
+    List includes gravatar, default avatar and any uploaded avatars.
     """
-    The part that's the least straightforward about views in this module is how they 
-    determine their redirects after they have finished computation.
-
-    In short, they will try and determine the next place to go in the following order:
-
-    1. If there is a variable named ``next`` in the *POST* parameters, the view will
-    redirect to that variable's value.
-    2. If there is a variable named ``next`` in the *GET* parameters, the view will
-    redirect to that variable's value.
-    3. If Django can determine the previous page from the HTTP headers, the view will
-    redirect to that previous page.
-    """
-    next = request.POST.get('next', request.GET.get('next',
-        request.META.get('HTTP_REFERER', None)))
-    if not next:
-        next = request.path
-    return next
-    
-def _notification_updated(request, avatar):
-    notification.send([request.user], "avatar_updated",
-        {"user": request.user, "avatar": avatar})
-    if friends:
-        notification.send((x['friend'] for x in
-                Friendship.objects.friends_for_user(request.user)),
-            "avatar_friend_updated",
-            {"user": request.user, "avatar": avatar}
-        )
-
-def _get_avatars(user):
-    # Default set. Needs to be sliced, but that's it. Keep the natural order.
+    #determine avatar data for the view
+    avatar_data = list()
     avatars = user.avatar_set.all()
-    
-    # Current avatar
-    primary_avatar = avatars.order_by('-primary')[:1]
-    if primary_avatar:
-        avatar = primary_avatar[0]
-    else:
-        avatar = None
-    
-    if AVATAR_MAX_AVATARS_PER_USER == 1:
-        avatars = primary_avatar
-    else:
-        # Slice the default set now that we used the queryset for the primary avatar
-        avatars = avatars[:AVATAR_MAX_AVATARS_PER_USER]
-    return (avatar, avatars)    
+    #iterate through uploaded avatars
+    for avatar in avatars:
+        datum = {
+            'id': avatar.id,
+            'avatar_type': 'uploaded_avatar',
+            'url': avatar.avatar_url(avatar_size),
+            'is_primary': avatar.primary
+        }
+        avatar_data.append(datum)
 
-@login_required
-@csrf.csrf_protect
-def add(request, extra_context=None, next_override=None,
-        upload_form=UploadAvatarForm, *args, **kwargs):
-    if extra_context is None:
-        extra_context = {}
-    avatar, avatars = _get_avatars(request.user)
-    upload_avatar_form = upload_form(request.POST or None,
-        request.FILES or None, user=request.user)
-    if request.method == "POST" and 'avatar' in request.FILES:
-        if upload_avatar_form.is_valid():
-            avatar = Avatar(
-                user = request.user,
-                primary = True,
-            )
+    #add gravatar datum
+    gravatar_datum = {
+        'avatar_type': 'gravatar',
+        'url': user.get_gravatar_url(avatar_size),
+        'is_primary': (user.avatar_type == 'g')
+    }
+    avatar_data.append(gravatar_datum)
+
+    #add default avatar datum
+    default_datum = {
+        'avatar_type': 'default_avatar',
+        'url': user.get_default_avatar_url(avatar_size),
+        'is_primary': (user.avatar_type == 'n')
+    }
+    avatar_data.append(default_datum)
+
+    #if there are >1 primary avatar, select just one
+    primary_avatars = filter(lambda v: v['is_primary'], avatar_data)
+    #force just one primary avatar if there are >1
+    if len(primary_avatars) > 1:
+        gravatars = filter(
+                        lambda v: v.type == 'gravatar',
+                        primary_avatars
+                    )
+
+        def clear_primary(datum):
+            datum['is_primary'] = False
+        map(clear_primary, primary_avatars)
+
+        if len(gravatars):
+            gravatars[0]['is_primary'] = True
+        else:
+            primary_avatars[0]['is_primary'] = True
+            
+    #insert primary avatar first
+    primary_avatars = filter(lambda v: v['is_primary'], avatar_data)
+    if len(primary_avatars):
+        primary_avatar = primary_avatars[0]
+        avatar_data.remove(primary_avatar)
+        avatar_data.insert(0, primary_avatar)
+
+    return avatar_data, bool(avatars.count())
+
+
+def redirect_to_show_list(user_id):
+    return HttpResponseRedirect(
+        reverse('askbot_avatar_show_list', kwargs={'user_id': user_id})
+    )
+   
+
+@admin_or_owner_required
+def show_list(request, user_id=None, extra_context=None, avatar_size=128):
+    """lists user's avatars, including gravatar and the default avatar"""
+    user = get_object_or_404(User, pk=user_id)
+    avatar_data, has_uploaded_avatar = get_avatar_data(user, avatar_size)
+    context = {
+        'avatar_data': avatar_data,
+        'has_uploaded_avatar': has_uploaded_avatar,
+        'max_avatars': avatar_settings.AVATAR_MAX_AVATARS_PER_USER,
+        'page_class': 'user-profile-page',
+        'upload_avatar_form': UploadAvatarForm(user=user),
+        'view_user': user
+    }
+    context.update(extra_context or {})
+    return render(request, 'avatar/show_list.html', context)
+
+@admin_or_owner_required
+def set_primary(request, user_id=None, extra_context=None, avatar_size=128):
+    """changes default uploaded avatar"""
+    user = get_object_or_404(User, pk=user_id)
+
+    if request.method == "POST":
+        updated = False
+        form = PrimaryAvatarForm(
+                        request.POST,
+                        user=user,
+                        avatars=user.avatar_set.all()
+                    )
+        if 'choice' in request.POST and form.is_valid():
+            avatar = Avatar.objects.get(id=form.cleaned_data['choice'])
+            avatar.primary = True
+            avatar.save()
+            avatar_updated.send(sender=Avatar, user=request.user, avatar=avatar)
+    return redirect_to_show_list(user_id)
+
+
+@admin_or_owner_required
+def upload(request, user_id=None):
+    user = get_object_or_404(User, pk=user_id)
+    if request.method == 'POST' and 'avatar' in request.FILES:
+        form = UploadAvatarForm(
+                        request.POST,
+                        request.FILES,
+                        user=user)
+        if form.is_valid():
+            avatar = Avatar(user=user, primary=True)
             image_file = request.FILES['avatar']
             avatar.avatar.save(image_file.name, image_file)
             avatar.save()
-            request.user.message_set.create(
-                message=_("Successfully uploaded a new avatar."))
-            if notification:
-                _notification_updated(request, avatar)
-            return HttpResponseRedirect(next_override or _get_next(request))
-    data = { 
-        'avatar': avatar, 
-        'avatars': avatars, 
-        'upload_avatar_form': upload_avatar_form,
-        'next': next_override or _get_next(request),
-        'view_user': request.user,
-        'page_class': 'avatar-page',
-    }
-    if extra_context:
-        data.update(extra_context)
+            avatar_updated.send(sender=Avatar, user=user, avatar=avatar)
 
-    return render(request, 'avatar/add.html', data)
+    return redirect_to_show_list(user_id)
 
-@login_required
-@csrf.csrf_protect
-def change(request, extra_context=None, next_override=None,
-        upload_form=UploadAvatarForm, primary_form=PrimaryAvatarForm,
-        *args, **kwargs):
-    if extra_context is None:
-        extra_context = {}
-    avatar, avatars = _get_avatars(request.user)
-    if avatar:
-        kwargs = {'initial': {'choice': avatar.id}}
-    else:
-        kwargs = {}
-    upload_avatar_form = upload_form(user=request.user, **kwargs)
-    primary_avatar_form = primary_form(request.POST or None,
-        user=request.user, avatars=avatars, **kwargs)
-    if request.method == "POST":
-        updated = False
-        if 'choice' in request.POST and primary_avatar_form.is_valid():
-            avatar = Avatar.objects.get(id=
-                primary_avatar_form.cleaned_data['choice'])
+
+def delete(request, avatar_id):
+    avatar = get_object_or_404(Avatar, pk=avatar_id)
+    user = request.user
+    if request.method == 'POST' \
+        and user.is_authenticated() \
+        and (user.is_administrator_or_moderator() \
+            or avatar.user_id == user.id):
+        avatar_type = avatar.user.avatar_type
+        avatar.delete()
+        avatar.user.avatar_type = avatar_type
+        avatar.user.save()
+        if avatar_type == 'g':
+            avatar.user.avatar_set.update(primary=False)
+
+    return redirect_to_show_list(avatar.user_id)
+
+
+@admin_or_owner_required
+def enable_gravatar(request, user_id=None):
+    if request.method == 'POST':
+        user = get_object_or_404(User, pk=user_id)
+        user.avatar_type = 'g'
+        user.save()
+        user.avatar_set.update(primary=False)
+    return redirect_to_show_list(user_id)
+
+
+@admin_or_owner_required
+def enable_default_avatar(request, user_id=None):
+    if request.method == 'POST':
+        user = get_object_or_404(User, pk=user_id)
+        user.avatar_type = 'n'
+        user.save()
+        user.avatar_set.update(primary=False)
+    return redirect_to_show_list(user_id)
+
+
+@admin_or_owner_required
+def disable_gravatar(request, user_id=None):
+    if request.method == 'POST':
+        user = get_object_or_404(User, pk=user_id)
+        user.avatar_type = 'a'
+        user.save()
+        if user.avatar_set.count():
+            avatar = user.avatar_set.all()[0]
             avatar.primary = True
             avatar.save()
-            updated = True
-            request.user.message_set.create(
-                message=_("Successfully updated your avatar."))
-        if updated and notification:
-            _notification_updated(request, avatar)
-        return HttpResponseRedirect(next_override or _get_next(request))
-    data = {
-        'avatar': avatar, 
-        'avatars': avatars,
-        'upload_avatar_form': upload_avatar_form,
-        'primary_avatar_form': primary_avatar_form,
-        'next': next_override or _get_next(request), 
-        'view_user': request.user,
-        'page_class': 'avatar-page',
-    }
-    if extra_context:
-        data.update(extra_context)
-
-    return render(request, 'avatar/change.html', data)
-
-@login_required
-@csrf.csrf_protect
-def delete(request, extra_context=None, next_override=None, *args, **kwargs):
-    if extra_context is None:
-        extra_context = {}
-    avatar, avatars = _get_avatars(request.user)
-    delete_avatar_form = DeleteAvatarForm(request.POST or None,
-        user=request.user, avatars=avatars)
-    if request.method == 'POST':
-        if delete_avatar_form.is_valid():
-            ids = delete_avatar_form.cleaned_data['choices']
-            if unicode(avatar.id) in ids and avatars.count() > len(ids):
-                # Find the next best avatar, and set it as the new primary
-                for a in avatars:
-                    if unicode(a.id) not in ids:
-                        a.primary = True
-                        a.save()
-                        if notification:
-                            _notification_updated(request, a)
-                        break
-            Avatar.objects.filter(id__in=ids).delete()
-            request.user.message_set.create(
-                message=_("Successfully deleted the requested avatars."))
-            return HttpResponseRedirect(next_override or _get_next(request))
-    data = {
-        'avatar': avatar, 
-        'avatars': avatars,
-        'delete_avatar_form': delete_avatar_form,
-        'next': next_override or _get_next(request),
-        'view_user': request.user,
-        'page_class': 'avatar-page',
-    }
-    if extra_context:
-        data.update(extra_context)
-
-    return render(request, 'avatar/confirm_delete.html', data)
-
-def render_primary(request, user_id = None, *args, **kwargs):
-    user = models.User.objects.get(id = user_id)
-    kwargs['user'] = user.username
-    return django_avatar_render_primary(request, *args, **kwargs)
+            avatar_updated.send(sender=Avatar, user=request.user, avatar=avatar)
+    return redirect_to_show_list(user_id)